// old/agir-work-planning-web/src/app/map/config/layers/filter-enum.ts
export enum FilterId {
  id = 'id',
  idGcaf = 'idGcaf',
  noGeoAccessory = 'noGeomatiqueAccessoire',
  noGeoValve = 'noGeomatiqueVanne',
  noGeoSegment = 'noGeomatiqueSegment',
  noGeoSewerManhole = 'noGeomatiqueRegard',
  noGeoChamber = 'noGeomatiqueChambre',
  noGeoJoin = 'noGeomatiqueRaccord',
  noGeoSump = 'noGeomatiquePuisard'
}